requirements.txt
click==6.7
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.6
django-localflavor==1.5.2
django-toolbelt==0.0.1
gunicorn==19.7.1
itsdangerous==0.24
MarkupSafe==0.23
plaid-python==2.0.2
psycopg2==2.7.3.1
pytz==2017.2
requests==2.13.0
static3==0.7.0
virtualenv==15.1.0
Werkzeug==0.11.15
whitenoise==3.3.1
